# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NexaTech Programming Learning Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  clerkId: string;
  email: string;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export interface UserUpdate {
  firstName?: string;
  lastName?: string;
}

export interface WeekSummary {
  weekId: number;
  weekNumber: number;
  title: string;
  lessonsCompleted: number;
  totalLessons: number;
  isUnlocked: boolean;
}

export interface UserStats {
  totalScore: number;
  lessonsCompleted: number;
  exercisesCompleted: number;
  currentWeek: number;
  currentStreak: number;
  totalLessons: number;
  totalExercises: number;
  weeklyProgress?: WeekSummary[];
}

export interface Week {
  id: number;
  weekNumber: number;
  title: string;
  description: string;
  objectives: string[];
  totalLessons: number;
  /** @nullable */
  imageUrl?: string | null;
}

export type LessonType = typeof LessonType[keyof typeof LessonType];


export const LessonType = {
  video: 'video',
  reading: 'reading',
  exercise: 'exercise',
  quiz: 'quiz',
} as const;

export interface Lesson {
  id: number;
  weekId: number;
  order: number;
  title: string;
  description: string;
  durationMinutes: number;
  type: LessonType;
  /** @nullable */
  videoUrl?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface WeekDetail {
  id: number;
  weekNumber: number;
  title: string;
  description: string;
  objectives: string[];
  /** @nullable */
  imageUrl?: string | null;
  lessons: Lesson[];
}

export type LessonDetailType = typeof LessonDetailType[keyof typeof LessonDetailType];


export const LessonDetailType = {
  video: 'video',
  reading: 'reading',
  exercise: 'exercise',
  quiz: 'quiz',
} as const;

export type ExerciseType = typeof ExerciseType[keyof typeof ExerciseType];


export const ExerciseType = {
  multiple_choice: 'multiple_choice',
  true_false: 'true_false',
  fill_blank: 'fill_blank',
  code_completion: 'code_completion',
} as const;

export type ExerciseDifficulty = typeof ExerciseDifficulty[keyof typeof ExerciseDifficulty];


export const ExerciseDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface Exercise {
  id: number;
  lessonId: number;
  weekId: number;
  order: number;
  question: string;
  type: ExerciseType;
  difficulty: ExerciseDifficulty;
  points: number;
  /** @nullable */
  options?: string[] | null;
  /** @nullable */
  hint?: string | null;
  /** @nullable */
  explanation?: string | null;
}

export interface LessonDetail {
  id: number;
  weekId: number;
  order: number;
  title: string;
  description: string;
  durationMinutes: number;
  type: LessonDetailType;
  /** @nullable */
  videoUrl?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  content: string;
  keyPoints?: string[];
  exercises?: Exercise[];
}

export interface ExerciseSubmission {
  answer: string;
}

export interface SubmissionResult {
  correct: boolean;
  pointsEarned: number;
  explanation: string;
  correctAnswer: string;
}

export interface LessonProgress {
  lessonId: number;
  userId: number;
  completed: boolean;
  score: number;
  /** @nullable */
  completedAt?: string | null;
  timeSpentMinutes?: number;
}

export interface LessonProgressUpdate {
  completed: boolean;
  timeSpentMinutes?: number;
}

export interface WeekProgress {
  weekId: number;
  lessonsCompleted: number;
  totalLessons: number;
  totalScore: number;
  isCompleted: boolean;
  lessonProgress?: LessonProgress[];
}

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  firstName: string;
  lastName: string;
  totalScore: number;
  lessonsCompleted: number;
  /** @nullable */
  avatarUrl: string | null;
}

export type GetExercisesParams = {
lessonId?: number;
weekId?: number;
difficulty?: GetExercisesDifficulty;
};

export type GetExercisesDifficulty = typeof GetExercisesDifficulty[keyof typeof GetExercisesDifficulty];


export const GetExercisesDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;