# lib/api-zod/src/generated/types/lesson.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NexaTech Programming Learning Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { LessonType } from './lessonType';

export interface Lesson {
  id: number;
  weekId: number;
  order: number;
  title: string;
  description: string;
  durationMinutes: number;
  type: LessonType;
  /** @nullable */
  videoUrl?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}