# lib/api-zod/src/generated/types/submissionResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NexaTech Programming Learning Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SubmissionResult {
  correct: boolean;
  pointsEarned: number;
  explanation: string;
  correctAnswer: string;
}