# lib/api-zod/src/generated/types/weekProgress.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NexaTech Programming Learning Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { LessonProgress } from './lessonProgress';

export interface WeekProgress {
  weekId: number;
  lessonsCompleted: number;
  totalLessons: number;
  totalScore: number;
  isCompleted: boolean;
  lessonProgress?: LessonProgress[];
}