# lib/api-zod/src/generated/types/weekSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NexaTech Programming Learning Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface WeekSummary {
  weekId: number;
  weekNumber: number;
  title: string;
  lessonsCompleted: number;
  totalLessons: number;
  isUnlocked: boolean;
}